refactor(#1697): split config.rs unit tests into config_tests.rs (#[path], epic #1135)
Mechanical move only — no test body changed. config.rs 1165 -> 789 lines,
under the ~800 source campsite target, which frees the budget the #1697
bridge needs. Follows the established `#[path = "*_tests.rs"]` pattern
(observability/catalog.rs, export/arrow_size.rs, storage/producer_fault.rs).

Refs #1697, #1135
